+++ CHANGELOG.md
@@ -130,6 +130,7 @@ Changelog
 - Added support for quickly adding cache header based on route match
 - Added support for quickly adding allow origin header based on route match
 - Added support for quickly re-routing to defined 404 handler
+- Added support for length based types (`length`, `shorter_than`, and `longer_than`)
 - Updated all default output formats to gracefully handle error dictionaries
 - Documentation generation was moved to API instances to enable easier customization
 - Now correctly identifies and handles custom encodings